Change node files from javascript to typescript format
