--- src/controllers/home.js
@@ -1,9 +1,0 @@
-class Home {
-  static mainAction(req, res) {
-    res.json({
-      status: true,
-    });
-  }
-}
-
-module.exports = Home;
