--- src/routes/index.js
@@ -1,12 +1,0 @@
-const HomeController = require('../controllers/home');
-
-class Routes {
-  static make(router) {
-    router.get(
-      '/',
-      HomeController.mainAction,
-    );
-  }
-}
-
-module.exports = Routes;
